requirements.txt
fpdf2==2.8.2
python_nmap==0.7.1
Requests==2.32.3
scapy==2.6.1
